requirements.txt
python-dotenv
Flask
watson_developer_cloud
